Лекция 2.5
Создан общий класс для методов тестов TestBase.
Из классов GroupCreationTests,
ContactCreationTest, GroupDeletionTest перемещены все методы, кроме тестов, в общий класс.
