+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/ContactCreationTest.java
@@ -1,95 +1,16 @@
 package ru.stqa.pft.addressbook;
 
-import org.openqa.selenium.*;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import java.util.concurrent.TimeUnit;
-
-public class ContactCreationTest {
-    private WebDriver wd;
-    private JavascriptExecutor js;
-
-    @BeforeMethod(alwaysRun = true)
-    public void setUp() throws Exception {
-        System.setProperty("webdriver.chrome.driver", "C:\\chromedriver\\chromedriver.exe");
-        wd = new ChromeDriver();
-        wd.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
-        js = (JavascriptExecutor) wd;
-        wd.get("http://localhost/addressbook/group.php");
-        login("admin", "secret");
-    }
-
-    private void login(String username, String password) {
-        wd.findElement(By.name("user")).click();
-        wd.findElement(By.name("user")).sendKeys(username);
-        wd.findElement(By.name("pass")).click();
-        wd.findElement(By.name("pass")).sendKeys(password);
-        wd.findElement(By.xpath("//input[@value='Login']")).click();
-    }
+public class ContactCreationTest extends TestBase {
 
     @Test
     public void testContactCreation() throws Exception {
         gotoContactPage();
-        initContactPage();
+        initContactCreation();
         fillContactForm(new ContactData("Ostap", "Bender", "+79998882211", "testTest@mail.ru"));
-        submit();
+        submitContactCreation();
         returnHomePage();
     }
 
-    private void returnHomePage() {
-        wd.findElement(By.linkText("home page")).click();
-    }
-
-    private void submit() {
-        wd.findElement(By.xpath("//div[@id='content']/form/input[21]")).click();
-    }
-
-    private void fillContactForm(ContactData contactData) {
-        wd.findElement(By.name("firstname")).sendKeys(contactData.getFirstname());
-        wd.findElement(By.name("lastname")).click();
-        wd.findElement(By.name("lastname")).sendKeys(contactData.getLastname());
-        wd.findElement(By.name("mobile")).click();
-        wd.findElement(By.name("mobile")).sendKeys(contactData.getMobile());
-        wd.findElement(By.name("email")).click();
-        wd.findElement(By.name("email")).sendKeys(contactData.getEmail());
-    }
-
-    private void initContactPage() {
-        wd.findElement(By.name("firstname")).click();
-    }
-
-    private void gotoContactPage() {
-        wd.findElement(By.linkText("add new")).click();
-    }
-
-    @AfterMethod(alwaysRun = true)
-    public void tearDown() throws Exception {
-        logout();
-        wd.quit();
-    }
-
-    private void logout() {
-        wd.findElement(By.linkText("Logout")).click();
-    }
-
-    private boolean isElementPresent(By by) {
-        try {
-            wd.findElement(by);
-            return true;
-        } catch (NoSuchElementException e) {
-            return false;
-        }
-    }
-
-    private boolean isAlertPresent() {
-        try {
-            wd.switchTo().alert();
-            return true;
-        } catch (NoAlertPresentException e) {
-            return false;
-        }
-    }
 }
