+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/GroupCreationTests.java
@@ -1,36 +1,8 @@
 package ru.stqa.pft.addressbook;
 
-import org.openqa.selenium.*;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import java.util.concurrent.TimeUnit;
-
-public class GroupCreationTests {
-    private WebDriver wd;
-    private JavascriptExecutor js;
-
-    @BeforeMethod(alwaysRun = true)
-    public void setUp() throws Exception {
-        System.setProperty("webdriver.chrome.driver", "C:\\chromedriver\\chromedriver.exe");
-        wd = new ChromeDriver();
-        wd.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
-        js = (JavascriptExecutor) wd;
-        wd.get("http://localhost/addressbook/group.php");
-        login("admin", "secret");
-    }
-
-    private void login(String username, String password) {
-        wd.findElement(By.name("user")).click();
-        wd.findElement(By.name("user")).clear();
-        wd.findElement(By.name("user")).sendKeys(username);
-        wd.findElement(By.name("pass")).click();
-        wd.findElement(By.name("pass")).clear();
-        wd.findElement(By.name("pass")).sendKeys(password);
-        wd.findElement(By.xpath("//input[@value='Login']")).click();
-    }
+public class GroupCreationTests extends TestBase {
 
     @Test
     public void testGroupCreation() throws Exception {
@@ -41,59 +13,4 @@ public class GroupCreationTests {
         returnGroupPage();
     }
 
-    private void returnGroupPage() {
-        wd.findElement(By.linkText("group page")).click();
-    }
-
-    private void submitGroupCreation() {
-        wd.findElement(By.name("submit")).click();
-    }
-
-    private void fillGroupForm(GroupData groupData) {
-        wd.findElement(By.name("group_name")).click();
-        wd.findElement(By.name("group_name")).clear();
-        wd.findElement(By.name("group_name")).sendKeys(groupData.getName());
-        wd.findElement(By.name("group_header")).click();
-        wd.findElement(By.name("group_header")).clear();
-        wd.findElement(By.name("group_header")).sendKeys(groupData.getHeader());
-        wd.findElement(By.name("group_footer")).click();
-        wd.findElement(By.name("group_footer")).clear();
-        wd.findElement(By.name("group_footer")).sendKeys(groupData.getFooter());
-    }
-
-    private void initGroupCreation() {
-        wd.findElement(By.name("new")).click();
-    }
-
-    private void gotoGroupPage() {
-        wd.findElement(By.linkText("groups")).click();
-    }
-
-    @AfterMethod(alwaysRun = true)
-    public void tearDown() throws Exception {
-        logout();
-        wd.quit();
-    }
-
-    private void logout() {
-        wd.findElement(By.linkText("Logout")).click();
-    }
-
-    private boolean isElementPresent(By by) {
-        try {
-            wd.findElement(by);
-            return true;
-        } catch (NoSuchElementException e) {
-            return false;
-        }
-    }
-
-    private boolean isAlertPresent() {
-        try {
-            wd.switchTo().alert();
-            return true;
-        } catch (NoAlertPresentException e) {
-            return false;
-        }
-    }
 }
