+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/GroupDeletionTest.java
@@ -1,62 +1,16 @@
 package ru.stqa.pft.addressbook;
 
-import org.openqa.selenium.*;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import java.util.concurrent.TimeUnit;
+public class GroupDeletionTest extends TestBase {
 
-public class GroupDeletionTest {
-    private WebDriver wd;
-
-    private JavascriptExecutor js;///
-
-    @BeforeMethod(alwaysRun = true)
-    public void setUp() throws Exception {
-        System.setProperty("webdriver.chrome.driver", "C:\\chromedriver\\chromedriver.exe");
-        wd = new ChromeDriver();
-        wd.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
-        js = (JavascriptExecutor) wd;
-    }
 
     @Test
     public void testGroupDeletion() throws Exception {
-        wd.get("http://localhost/addressbook/");
-        wd.findElement(By.name("user")).click();
-        wd.findElement(By.name("user")).clear();
-        wd.findElement(By.name("user")).sendKeys("admin");
-        wd.findElement(By.id("LoginForm")).click();
-        wd.findElement(By.name("pass")).clear();
-        wd.findElement(By.name("pass")).sendKeys("secret");
-        wd.findElement(By.xpath("//input[@value='Login']")).click();
-        wd.findElement(By.linkText("groups")).click();
-        wd.findElement(By.name("selected[]")).click();
-        wd.findElement(By.name("delete")).click();
-        wd.findElement(By.linkText("group page")).click();
+        gotoGroupPage();
+        selectGroup();
+        deleteSelectedGroup();
+        returnGroupPage();
     }
 
-    @AfterMethod(alwaysRun = true)
-    public void tearDown() throws Exception {
-        wd.quit();
-    }
-
-    private boolean isElementPresent(By by) {
-        try {
-            wd.findElement(by);
-            return true;
-        } catch (NoSuchElementException e) {
-            return false;
-        }
-    }
-
-    private boolean isAlertPresent() {
-        try {
-            wd.switchTo().alert();
-            return true;
-        } catch (NoAlertPresentException e) {
-            return false;
-        }
-    }
 }
