+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/TestBase.java
@@ -1,0 +1,118 @@
+package ru.stqa.pft.addressbook;
+
+import org.openqa.selenium.*;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import java.util.concurrent.TimeUnit;
+
+public class TestBase {
+    private WebDriver wd;
+    private JavascriptExecutor js;
+
+    @BeforeMethod(alwaysRun = true)
+    public void setUp() throws Exception {
+        System.setProperty("webdriver.chrome.driver", "C:\\chromedriver\\chromedriver.exe");
+        wd = new ChromeDriver();
+        wd.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
+        js = (JavascriptExecutor) wd;
+        wd.get("http://localhost/addressbook/group.php");
+        login("admin", "secret");
+    }
+
+    private void login(String username, String password) {
+        wd.findElement(By.name("user")).click();
+        wd.findElement(By.name("user")).clear();
+        wd.findElement(By.name("user")).sendKeys(username);
+        wd.findElement(By.name("pass")).click();
+        wd.findElement(By.name("pass")).clear();
+        wd.findElement(By.name("pass")).sendKeys(password);
+        wd.findElement(By.xpath("//input[@value='Login']")).click();
+    }
+
+    protected void returnGroupPage() {
+        wd.findElement(By.linkText("group page")).click();
+    }
+
+    protected void submitGroupCreation() {
+        wd.findElement(By.name("submit")).click();
+    }
+
+    protected void fillGroupForm(GroupData groupData) {
+        wd.findElement(By.name("group_name")).click();
+        wd.findElement(By.name("group_name")).clear();
+        wd.findElement(By.name("group_name")).sendKeys(groupData.getName());
+        wd.findElement(By.name("group_header")).click();
+        wd.findElement(By.name("group_header")).clear();
+        wd.findElement(By.name("group_header")).sendKeys(groupData.getHeader());
+        wd.findElement(By.name("group_footer")).click();
+        wd.findElement(By.name("group_footer")).clear();
+        wd.findElement(By.name("group_footer")).sendKeys(groupData.getFooter());
+    }
+
+    protected void initGroupCreation() {
+        wd.findElement(By.name("new")).click();
+    }
+
+    protected void gotoGroupPage() {
+        wd.findElement(By.linkText("groups")).click();
+    }
+
+    @AfterMethod(alwaysRun = true)
+    public void tearDown() throws Exception {
+        wd.quit();
+    }
+
+    private boolean isElementPresent(By by) {
+        try {
+            wd.findElement(by);
+            return true;
+        } catch (NoSuchElementException e) {
+            return false;
+        }
+    }
+
+    private boolean isAlertPresent() {
+        try {
+            wd.switchTo().alert();
+            return true;
+        } catch (NoAlertPresentException e) {
+            return false;
+        }
+    }
+
+    protected void returnHomePage() {
+        wd.findElement(By.linkText("home page")).click();
+    }
+
+    protected void submitContactCreation() {
+        wd.findElement(By.xpath("//div[@id='content']/form/input[21]")).click();
+    }
+
+    protected void fillContactForm(ContactData contactData) {
+        wd.findElement(By.name("firstname")).sendKeys(contactData.getFirstname());
+        wd.findElement(By.name("lastname")).click();
+        wd.findElement(By.name("lastname")).sendKeys(contactData.getLastname());
+        wd.findElement(By.name("mobile")).click();
+        wd.findElement(By.name("mobile")).sendKeys(contactData.getMobile());
+        wd.findElement(By.name("email")).click();
+        wd.findElement(By.name("email")).sendKeys(contactData.getEmail());
+    }
+
+    protected void initContactCreation() {
+        wd.findElement(By.name("firstname")).click();
+    }
+
+    protected void gotoContactPage() {
+        wd.findElement(By.linkText("add new")).click();
+    }
+
+    protected void deleteSelectedGroup() {
+        wd.findElement(By.name("delete")).click();
+    }
+
+    protected void selectGroup() {
+        wd.findElement(By.name("selected[]")).click();
+    }
+}
